Added locale configuration to environmet
This change allows the app to dynamically select the appropriate images and videos by concatenating the file name.

This commit also moves the list of regions from the 'localization screen' to the environment configuration. Ideally this would be API driven but at least it's more configurable here than in the code.