Quick fix to address word wrapping issue.
If a string begins with a space (or is sliced in a way that puts a space
in the first position) then the splitTextLines() function would end up
in an infinite loop. Quick fix is to detect this case and remove the
space character from the front of the string.